# src/CoursePlatform.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CoursePlatform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/lessons": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "summary": "Creates a new lesson",
        "requestBody": {
          "description": "Lesson creation details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lesson created successfully"
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{id}": {
      "delete": {
        "tags": [
          "Lessons"
        ],
        "summary": "Deletes a lesson (Soft Delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lesson ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Lesson deleted successfully"
          },
          "404": {
            "description": "Lesson not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lessons"
        ],
        "summary": "Updates a lesson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lesson ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Update details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Lesson updated successfully"
          },
          "404": {
            "description": "Lesson not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/reorder": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "summary": "Reorders lessons within a course",
        "requestBody": {
          "description": "Reorder details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderLessonsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderLessonsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderLessonsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Lessons reordered successfully"
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ReorderLessonsDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "lessonIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}